requirements.txt
django
djangorestframework
django-cors-headers
django-phonenumber-field
Pillow